refactor: nutlink
